+++ src/imslim/compressors/__init__.py
@@ -1,5 +1,6 @@
 from .avif_compressor import AVIFCompressor
 from .jpeg_compressor import JPEGCompressor
+from .jxl_compressor import JXLCompressor
 from .png_compressor import PNGCompressor
 from .svg_compressor import SVGCompressor
 from .webp_compressor import WEBPCompressor
@@ -7,6 +8,7 @@ from .webp_compressor import WEBPCompressor
 __all__ = [
     "AVIFCompressor",
     "JPEGCompressor",
+    "JXLCompressor",
     "PNGCompressor",
     "SVGCompressor",
     "WEBPCompressor",